feat: unique test check health of server
