+++ server_tests/test_cases/video_generation_eval_test.py
@@ -121,56 +121,6 @@ class VideoGenerationEvalsTest(BaseTest):
             "eval_results": results,
         }
 
-    def _wait_for_server_ready(
-        self,
-        service_port: int = 8000,
-        max_attempts: int = 230,
-        retry_delay: int = 10,
-    ) -> bool:
-        """Wait for server to be ready using simple HTTP health check.
-
-        Args:
-            service_port: Port where the server is running.
-            max_attempts: Maximum number of retry attempts.
-            retry_delay: Seconds to wait between retries.
-
-        Returns:
-            bool: True if server is ready, False otherwise.
-        """
-        logger.info("Waiting for server to be ready...")
-        health_url = f"http://localhost:{service_port}/tt-liveness"
-        logger.info("Health URL: %s", health_url)
-
-        for attempt in range(1, max_attempts + 1):
-            try:
-                response = requests.get(health_url, timeout=10)
-                if response.status_code == 200:
-                    data = response.json()
-                    if data.get("status") == "alive" and data.get("model_ready"):
-                        logger.info(
-                            "Server is ready after %s attempt(s)",
-                            attempt,
-                        )
-                        return True
-                logger.info(
-                    "Server not ready (attempt %s/%s), retrying in %ss...",
-                    attempt,
-                    max_attempts,
-                    retry_delay,
-                )
-            except requests.exceptions.RequestException as e:
-                logger.info(
-                    "Health check failed (attempt %s/%s): %s, retrying in %ss...",
-                    attempt,
-                    max_attempts,
-                    e,
-                    retry_delay,
-                )
-            time.sleep(retry_delay)
-
-        logger.error("Server health check failed after %s attempts", max_attempts)
-        return False
-
     def _generate_videos(
         self,
         prompts: list[str],
@@ -197,7 +147,7 @@ class VideoGenerationEvalsTest(BaseTest):
             shutil.rmtree(output_dir)
         output_dir.mkdir(parents=True, exist_ok=True)
 
-        if not self._wait_for_server_ready():
+        if not self.wait_for_server_ready():
             raise RuntimeError("Server health check failed - server not ready")
 
         base_url = server_url or SERVER_BASE_URL
